requirements.txt
toml
python-magic